+++ @utahdts/utah-design-system/react/components/containers/tabs/TabGroup.jsx
@@ -1,4 +1,4 @@
-import { useCallback, useId, useMemo, useRef } from 'react';
+import { useCallback, useEffect, useId, useMemo, useRef } from 'react';
 import { useImmer } from 'use-immer';
 import { joinClassNames } from '../../../util/joinClassNames';
 import { TabGroupContext } from './context/TabGroupContext';
@@ -67,6 +67,18 @@ export function TabGroup({
       });
     },
     []
+  );
+
+  // check if "controlled" component's value changed
+  useEffect(
+    () => {
+      if (value !== undefined) {
+        setTabGroupState((draftState) => {
+          draftState.selectedTabId = value;
+        });
+      }
+    },
+    [value]
   );
 
   /** @type {TabGroupContextValue} */