+++ utah-design-system-website/src/react/components/websiteContent/library/components/buttons/button/ButtonExampleCodeReact.jsx
@@ -1,4 +1,4 @@
-import { formElementSizesEnum } from '@utahdts/utah-design-system';
+import { formElementSizesEnum, joinClassNames } from '@utahdts/utah-design-system';
 import { ExampleCodeReactProp } from '../../../../../sandbox/ExampleCodeReactProp';
 import { SandboxIndent } from '../../../../../sandbox/SandboxIndent';
 
@@ -32,8 +32,32 @@ export function ButtonExampleCodeReact({
       <br />
       <ExampleCodeReactProp displayProp={appearance ? `appearance="${appearance}"` : null} indentLevel={1} />
       <ExampleCodeReactProp displayProp={color ? `color="${color}"` : null} indentLevel={1} />
-      <ExampleCodeReactProp displayProp={(iconLeft !== 'none' && iconLeft) ? `iconLeft={Icons.${iconLeft}()}` : null} indentLevel={1} />
-      <ExampleCodeReactProp displayProp={(iconRight !== 'none' && iconRight) ? `iconRight={Icons.${iconRight}()}` : null} indentLevel={1} />
+      <ExampleCodeReactProp
+        displayProp={
+          ((iconLeft === 'none') || !iconLeft)
+            ? null
+            : (
+              `iconLeft={<span
+    className="${joinClassNames(iconLeft === 'IconChevron' ? 'utds-icon-after-chevron-down' : 'utds-icon-after-arrow-left', 'icon')}"
+    aria-hidden="true"
+  />}`
+            )
+        }
+        indentLevel={1}
+      />
+      <ExampleCodeReactProp
+        displayProp={
+          ((iconRight === 'none') || !iconRight)
+            ? null
+            : (
+              `iconRight={<span
+    className="${joinClassNames(iconRight === 'IconChevron' ? 'utds-icon-after-chevron-down' : 'utds-icon-after-arrow-right', 'icon')}"
+    aria-hidden="true"
+  />}`
+            )
+        }
+        indentLevel={1}
+      />
       <ExampleCodeReactProp displayProp={isBusy ? 'isBusy' : null} indentLevel={1} />
       <ExampleCodeReactProp displayProp={isDisabled ? 'isDisabled' : null} indentLevel={1} />
       <ExampleCodeReactProp displayProp={id ? `id="${id}"` : null} indentLevel={1} />