+++ utah-design-system-website/src/react/components/websiteContent/library/components/forms/Switch/SwitchExampleCodeReact.jsx
@@ -1,3 +1,4 @@
+import { joinClassNames } from '@utahdts/utah-design-system';
 import { ExampleCodeReactProp } from '../../../../../sandbox/ExampleCodeReactProp';
 
 /** @typedef {import('utah-design-system-website').SwitchExamplePropsShape} SwitchExamplePropsShape */
@@ -40,7 +41,19 @@ export function SwitchExampleCodeReact({
       <ExampleCodeReactProp displayProp={`value={${value}}`} indentLevel={1} />
       <ExampleCodeReactProp displayProp={size ? `size="${size}"` : null} indentLevel={1} />
       <ExampleCodeReactProp displayProp={width ? `width={${width}}` : null} indentLevel={1} />
-      <ExampleCodeReactProp displayProp={(icon && icon !== 'none') ? `sliderChildren={icons.${icon}()}` : null} indentLevel={1} />
+      <ExampleCodeReactProp
+        displayProp={
+          ((icon === 'none') || !icon)
+            ? null
+            : (
+              `sliderChildren={<span
+    className="${joinClassNames(icon === 'checkbox' ? 'utds-icon-after-checkbox' : 'utds-icon-after-visibility', 'icon')}"
+    aria-hidden="true"
+  />}`
+            )
+        }
+        indentLevel={1}
+      />
       /&gt;
     </>
   );