+++ utah-design-system-website/src/react/components/websiteContent/library/components/widgetsIndicators/badges/BadgesExampleCodeReact.jsx
@@ -20,16 +20,26 @@ export function BadgesExampleCodeReact({
 }) {
   return (
     <>
-      &lt;Badge
-      <br />
-      <ExampleCodeReactProp displayProp={className ? `className="${className}"` : null} indentLevel={1} />
-      <ExampleCodeReactProp displayProp={`title="${title}"`} indentLevel={1} />
-      &gt;
+      &lt;button className=&quot;button button--solid&quot; type=&quot;button&quot; &gt;
       <br />
       <SandboxIndent indentLevel={1} />
+      Button
+      <br />
+      <SandboxIndent indentLevel={1} />
+      &lt;Badge
+      <br />
+      <ExampleCodeReactProp displayProp={className ? `className="${className}"` : null} indentLevel={2} />
+      <ExampleCodeReactProp displayProp={`title="${title}"`} indentLevel={2} />
+      <SandboxIndent indentLevel={1} />
+      &gt;
+      <br />
+      <SandboxIndent indentLevel={2} />
       {children}
       <br />
+      <SandboxIndent indentLevel={1} />
       &lt;/Badge&gt;
+      <br />
+      &lt;/button&gt;
     </>
   );
 }