+++ CHANGELOG.md
@@ -6,12 +6,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [x.x.x] x/x/xxxx
+## Breaking Change
+- **[Breaking]** Remove FormContext
+  - Individual components inside a form now need individual onChange/value props
+  - `<Form state={state} setState={setState}><TextInput /></Form>`
+  - becomes ==>
+  - `<Form><TextInput onChange={(e) => {setState(...}} value={state.textInputValue} /></Form>`
+
 ## Added
 - Add header setting for `titleFunction` for single page apps
 
+# [2.0.3] 9/9/2024
+## Fixed
+- Make TabGroup respond to controlled changes
+
 # [2.0.2] 8/13/2024
 ## Fixed
-- Utah Header should have a white background color (just not on the main .utah-design-system css class).
+- Utah Header should have a white background color (just not on the main .utah-design-system css class)
 
 # [2.0.1] 8/5/2024
 ## Fixed